WIP refactoring to support replication
