+++ .idea/go.imports.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="GoImports">
+    <option name="excludedPackages">
+      <array>
+        <option value="golang.org/x/tools/go/ssa/interp/testdata/src/strings" />
+      </array>
+    </option>
+  </component>
+</project>
